// src/enums/InjuryType.ts
export enum InjuryType {
  // Head and Neck
  Concussion = "Concussion",
  NeckStrain = "Neck Strain",
  Stinger = "Stinger",

  // Upper Body
  ShoulderDislocation = "Shoulder Dislocation",
  RotatorCuffTear = "Rotator Cuff Tear",
  BrokenCollarbone = "Broken Collarbone",
  ACJointSprain = "AC Joint Sprain",
  ElbowSprain = "Elbow Sprain",
  BrokenArm = "Broken Arm",
  WristFracture = "Wrist Fracture",
  DislocatedFinger = "Dislocated Finger",
  BrokenFinger = "Broken Finger",
  ThumbSprain = "Thumb Sprain",

  // Core
  RibFracture = "Rib Fracture",
  AbdominalStrain = "Abdominal Strain",
  Hernia = "Hernia",
  LowerBackStrain = "Lower Back Strain",

  // Lower Body
  HipPointer = "Hip Pointer",
  GroinStrain = "Groin Strain",
  HamstringStrain = "Hamstring Strain",
  QuadStrain = "Quad Strain",
  KneeSprain = "Knee Sprain",
  ACLTear = "ACL Tear",
  MCLTear = "MCL Tear",
  LCLTear = "LCL Tear",
  PCLTear = "PCL Tear",
  MeniscusTear = "Meniscus Tear",
  PatellarTendinitis = "Patellar Tendinitis",
  AnkleSprain = "Ankle Sprain",
  HighAnkleSprain = "High Ankle Sprain",
  AchillesTear = "Achilles Tear",
  TurfToe = "Turf Toe",
  BrokenFoot = "Broken Foot",
  ToeFracture = "Toe Fracture",

  // General
  Contusion = "Contusion",
  Cramping = "Cramping",
  Fatigue = "Fatigue",
  Dehydration = "Dehydration",
}